Refactor code to break out spawning routine
